+++ src/core/platform-helpers/platform-helpers.h
@@ -104,6 +104,8 @@ public:
 	virtual void onRecordingPaused () const = 0;
 	virtual void stopRinging () const = 0;
 
+	virtual void setDeviceRotation (int orientation) const = 0;
+
 protected:
 	inline explicit PlatformHelpers (std::shared_ptr<LinphonePrivate::Core> core) : CoreAccessor(core) {}
 
@@ -170,6 +172,7 @@ public:
 	void onRecordingStarted () const override {};
 	void onRecordingPaused () const override {};
 	void stopRinging () const override {};
+	void setDeviceRotation (int orientation) const override {};
 
 protected:
 	bool checkIpAddressChanged();